pyproject.toml
[project]
name = "finance-metrics"
version = "0.1.0"
description = "Financial and economic metrics for retail demand forecasting"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yfinance>=0.2.66",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "fredapi>=0.5.2",
    "python-dotenv>=1.2.1",
    "lxml>=6.0.2",
    "html5lib>=1.1",
    "beautifulsoup4>=4.14.3",
    "selenium>=4.38.0",
    "playwright>=1.56.0",
    "pyarrow>=22.0.0",
    "boto3>=1.42.6",
    "polygon-api-client>=1.16.3",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
